+++ src/ipc/utils/get_model_client.ts
@@ -15,10 +15,8 @@ import type {
 } from "../../lib/schemas";
 import { getEnvVar } from "./read_env";
 import log from "electron-log";
-import {
-  FREE_OPENROUTER_MODEL_NAMES,
-  getLanguageModelProviders,
-} from "../shared/language_model_helpers";
+import { FREE_OPENROUTER_MODEL_NAMES } from "../shared/language_model_constants";
+import { getLanguageModelProviders } from "../shared/language_model_helpers";
 import { LanguageModelProvider } from "../ipc_types";
 import { createDyadEngine } from "./llm_engine_provider";
